+++ src/types.ts
@@ -4,3 +4,5 @@ export type CalendarValues = {
   start?: CalendarDate
   end?: CalendarDate
 }
+
+export type Buttons = ({ onClick }: { onClick: () => void }) => JSX.Element